Converted Detail and AccountHistory ids Integer -> BigInteger

+++ model/src/main/java/com/coinbase/exchange/model/Detail.java
@@ -1,11 +1,13 @@
 package com.coinbase.exchange.model;
 
+import java.math.BigInteger;
+
 /**
  * Created by irufus on 2/25/15.
  */
 public class Detail {
     private String order_id;
-    private Integer trade_id;
+    private BigInteger trade_id;
     private String product_id;
 
     public String getOrder_id() {
@@ -16,11 +18,11 @@ public class Detail {
         this.order_id = order_id;
     }
 
-    public Integer getTrade_id() {
+    public BigInteger getTrade_id() {
         return trade_id;
     }
 
-    public void setTrade_id(Integer trade_id) {
+    public void setTrade_id(BigInteger trade_id) {
         this.trade_id = trade_id;
     }
